Add files via upload
It isn't done completely yet, but I figured I should commit what I have now while I finish. I have it to where it sorts the data, partitions it, calls merge, sorts both partitions by the other dimension to get the other median, and adds the S11 partition and ignores the S22 partition since it is always dominated. I just have to code the comparisons of S11 to S21 and S12 and add those not dominated.